Cập nhật code mới và thay thế code cũ

+++ NMproject/NMproject/src/main/java/com/example/NMproject/dto/AccountResponse.java
@@ -8,24 +8,24 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AccountResponse {
-	private Integer userID;
+	private long userID;
 	private String username;
 	private String email;
 	private String name;
 	private String phone;
 	private String address;
 	private String imageLink = "/hinh_anh/avatar.jpg"; // Đặt giá trị mặc định cho avatarUrl
-	private int userRole; // Thêm userRole vào
+	private long userRole; // Thêm userRole vào
 
 	// Constructor không có avatarUrl và userRole
-	public AccountResponse(Integer userID, String email, String username) {
+	public AccountResponse(long userID, String email, String username) {
 		this.userID = userID;
 		this.email = email;
 		this.username = username;
 	}
 
 	// Constructor không có userRole
-	public AccountResponse(Integer userID, String username, String email, String name, String phone, String address) {
+	public AccountResponse(long userID, String username, String email, String name, String phone, String address) {
 		this.userID = userID;
 		this.username = username;
 		this.email = email;
@@ -35,8 +35,8 @@ public class AccountResponse {
 	}
 
 	// Constructor có đầy đủ tất cả các trường
-	public AccountResponse(Integer userID, String username, String email, String name, String phone, String address,
-			int userRole) {
+	public AccountResponse(long userID, String username, String email, String name, String phone, String address,
+			long userRole) {
 		this.userID = userID;
 		this.username = username;
 		this.email = email;
